Tidy new variables list component
